fix: properly respawn players on checkpoints
Previously, players would respawn on exact checkpoint location, which
could be problematic if said checkpoint origin is located in a prop
(meaning player would respawn in said prop and get stuck there).

To fix that, player location is saved each time they reach a checkpoint,
and this location is used to respawn them in case they die.